refactor: split tests for lhs and rhs
